Adds LOCAL_CLIENT check to Media Cloudinary
While dogfooding, I encountered an error message when implementing the Cloudinary MediaStore locally (clientId is `null`).

Fortunately, I knew this was being triggered by the `authorized` method and just had it return `true`.  But, I thought it might be best to only short circuit `authorized` when local (and preserve the test otherwise).