1116 - Updates docs for TinaCMS Context
We've changed `/admin` for both `tina-cloud-starter` and `tinacms init` and wanted to update the documentation to reflect that.

`/admin` will now take the user to a log in page.  `/exit-admin` has been replaced by `/admin/logout`.

This new page is also an easy gateway to using `TinaAdmin` should a user be interested in doing that (but I chose to leave that out of the docs for now).