💄Book a Demo Modal (#1860)
As per Levi's email **Re: Tina.io - Meet with Scott requires update**

Hi Josh,
⚠️ IMPORTANT please give this priority as Scott is on leave with no
signal. i.e. we might miss a demo because it’s in Scott’s calendar.
 
Here is my booking link:
[https://meetings-eu1.hubspot.com/levi-jackson](https://aus01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fmeetings-eu1.hubspot.com%2Flevi-jackson&data=05%7C02%7CJoshBerman%40ssw.com.au%7Cb0182aba987145e0971008dc906664ff%7Cac2f7c34b93548e9abdc11e5d4fcb2b0%7C0%7C0%7C638544016674491533%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C0%7C%7C%7C&sdata=5vZmfttkNLFYKAy8EY9gGSRjL%2Fz1Zg00c0ZE4j1K9Uk%3D&reserved=0)
 
1. Easy fix - Please investigate how Scott’s HubSpot booking link works
on the website and change the URL from Scott’s to mine
 
2. A better fix – Remember that most customers are from America and so
they generally would want to arrange a meeting within their work hours
(aka with Scott).
Therefore, it would be great to give the users a choice based on the
timezone. i.e. be able to select either Scott, Jean or I based on
location/timezone.

We have gone for `2`, a _better fix_:
- Added a modal for booking options when users click 'book a demo' 
- Added links to hubspots for Scott and Levi **❗AWAITING JEANS**

_✅all test passed by @bettybondoc_

![Screenshot 2024-06-20 at 11 53
14 am](https://github.com/tinacms/tina.io/assets/137844305/659e215a-e749-4140-a9e2-e05fc2fd7263)
**Figure: Laptop L view of new modal**

---------

Co-authored-by: joshua berman <jber4409@uni.sydney.edu.au>