Auto Translation - Adding Manual Trigger (#3589)
Fixed #3588 


To streamline this process and avoid creating numerous PRs for each
broken pipeline, I'm implementing a manual trigger for the translation
flow. This trigger would accept the pull request number as an input,
allowing you to re-translate a specific page directly without needing to
merge a new PR. This enhancement will significantly ease the burden of
re-triggering translations and improve the overall efficiency of the
workflow.